refactor business logic
