+++ src/main/java/org/example/homework/PrintCurrency/CurrencyISO.java
@@ -1,0 +1,19 @@
+package org.example.homework.PrintCurrency;
+
+public enum CurrencyISO {
+    RUB(0),
+    USD(1);
+
+
+    private final int arraysIndex;
+
+    CurrencyISO(int arraysIndex) {
+        this.arraysIndex = arraysIndex;
+    }
+
+    public int getArraysIndex() {
+        return this.arraysIndex;
+    }
+
+
+}
