+++ src/main/java/org/example/homework/PrintCurrency/PrintCurrency.java
@@ -1,24 +1,27 @@
 package org.example.homework.PrintCurrency;
 
+import org.example.homework.oop.ConsoleUserOutput;
+import org.example.homework.oop.UserOutput;
+
 public class PrintCurrency {
     private final CurrencyInput currencyInput;
+    private static UserOutput userOutput;
 
 
-    public PrintCurrency(CurrencyInput currencyInput) {
+    public PrintCurrency(CurrencyInput currencyInput, UserOutput userOutput) {
         this.currencyInput = currencyInput;
+        this.userOutput = userOutput;
     }
 
     public static void main(String[] args) {
 
-
         RubleCurrencyInput rubleCurrencyInput = new RubleCurrencyInput();
-        PrintCurrency printCurrency = new PrintCurrency(rubleCurrencyInput);
-        int amount = printCurrency.currencyInput.readAmount();
+        PrintCurrency printCurrency = new PrintCurrency(rubleCurrencyInput, new ConsoleUserOutput());
+        int amountInput = printCurrency.currencyInput.readAmount();
 
+        Amount amount = new Amount(amountInput);
 
-        Currency currency = new Currency(amount);
-
-        System.out.println(amount + " " + currency.displayAmount());
+        userOutput.displayText(amount.displayAmount());
 
     }
